requirements.txt
boto3==1.5.22
botocore==1.8.36
Django==1.11.5
django-admin-honeypot==1.0.0
django-debug-toolbar==1.9.1
django-health-check==3.5.1
django-hitcount==1.3.0
django-markdown2==0.3.1
django-markdownx==2.0.23
django-storages==1.6.5
django-widget-tweaks==1.4.1
docutils==0.14
jmespath==0.9.3
Markdown==2.6.11
markdown2==2.3.5
numpy==1.14.0
Pillow==5.0.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
sqlparse==0.2.4
